delete old Trace MH, update test suite
