+++ test/Spec.hs
@@ -78,7 +78,6 @@ main = hspec $ do
       TestInference.check_prior_trans `shouldBe` True
     it "Trace MH leaves posterior invariant" $ do
       TestInference.check_trace_trans `shouldBe` True
-      TestInference.check_mh_trans `shouldBe` True
     it "Trace MH leaves posterior invariant when the model has shifting support" $ do
       TestInference.check_trace_support `shouldBe` True
     -- too large to execute
